Refactored tests of conditional middleware
